Add checks linalg_eig_backward
This PR adds checks for the backward of `linalg.eig`, similar to those
deduced in https://github.com/pytorch/pytorch/pull/70253

It also modifies the function so that it does not save the input matrix,
as it's not necessary.

It also corrects the forward AD formula for it to be correct. Now all
the tests pass for `linalg.eig` and `linalg.eigvals`.

It also updates the docs to reflect better what's going on here.

ghstack-source-id: c93fd8153071b4598d2f506b06f3c597bb2d4f46
Pull Request resolved: https://github.com/pytorch/pytorch/pull/70527